refactor: Share connection among sinks
